# Conflicts:
#	src/main/java/keeper/project/homepage/controller/PostingController.java
#	src/main/java/keeper/project/homepage/service/PostingService.java
#	src/test/java/keeper/project/homepage/controller/CommentControllerTest.java
#	src/test/java/keeper/project/homepage/controller/PostingControllerTest.java
#	src/test/java/keeper/project/homepage/repository/CommentRepositoryTest.java
#	src/test/java/keeper/project/homepage/service/CommentServiceTest.java

+++ src/main/java/keeper/project/homepage/repository/member/MemberRankRepository.java
@@ -3,6 +3,6 @@ package keeper.project.homepage.repository.member;
 import keeper.project.homepage.entity.member.MemberRankEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface MemberRankRepository extends JpaRepository<MemberRankEntity, Integer> {
+public interface MemberRankRepository extends JpaRepository<MemberRankEntity, Long> {
 
 }
